solrXMLPEPWebLoad fixes for problems when switched to mySQL
I switched solrXMLPEPWebLoad to use mysql rather than SQLite because SQLite did not allow for concurrent runs updating the Solr database. In doing so, I've also made the tracking specific to the SOLRURL, since I may update databases on various URLs from my local PC.

The first mySQL conversion was problematic because I was using float in the schema for the mdate fields, and I needed to use decimal.