term counts
Got term counts working again from PEPEasy isolating Searchanalysis from Search in the process.  One issue was on PEPEasy itself, it was repeatedly sending requests for searchanalysis.  It may not have been necessary to isolate SearchAnalysis, but it helped with debugging and it's easier to follow and update the code when its separated.  The downside is that these functions (also morelike this) have so many parameters!