XSLT support for videos
Set up the conversion from the video file callout to the code required to display vdieos from Wistia.  We have several options here and there's a new option which worked best.  (I left a few other options as commented code in case we want to experiment later.  (The XSLT file is not finalized. )