S3 support,  Queryhelper fixes, Test updates
Merge of a few prior commits that were rebased away.    S3 works now for images.  S3 works but isn't' a good match for downloads, because it doesn't trigger the browser download process and the users choice of folder.  Need to work on that.  Fixes to queryhelper to fix issues with searchAnalysis when no boolean connector is specified.  Added prefix switches (experimental) to call out special query functions.  Fixes to tests, so all run successfully.