Experimental Updates - First try - new core for paras
First, this version removes Python 2 compatibility code.  We are solidly Python 3 now.

Then it adds experimental changes (not all clean coding) to see how we can best handle query within para.  This first versiion involves adding another core with all paras, so that we can use a join from the main Docs core in the filter query field to do the paragraph query, restricting the set of documents to those that match for any "main query".  The changes in solrPEPWebLoad load the new core.  No changes to implement the filter query are in the other modules: all testing was done in Solr Admin.  The rest of the changes here are the new core definition.

Unfortnately, I could not get highlighting in the main core to work when the filter query was active.

However, the next experiment is to add the paragraphs directly to the docs core!  That way, rather than a join, we might be able to take advantage of the multifield search and preserve highlighting!  But that's no in this build.

Also added a top level bat file to make it easier to update all the schemas without going down to the data level.