Fixed journal logo and query analysis
Fixed journal logo case of logo suffix to Logo to match files in xslt file.  Also, there was a .logo where sent back to the client which is wrong.  Added db url for admins in status.  Fixes to query analysis and terms.  Fixes to solrXMLPEPWebLoad programs to allow tunneling for remote SQL server (or Bitnami SQL install).