testing, fixes, update two endpoint's query params and update documentation
Changed timestamp to file_last_modified in get_what's new so it's not just when the file is uploaded, but rather when it's edited (processed)

updated endpoint documentation

Added code to search_text so (abstract_xml, summaries_xml, and text_xml) are only returned when abstract_requested is set, and only text_xml from that set is returned when just full_text_requested is set in the call to search_text.

added parameters required for PEP-Web to most_cited and most_downloaded.