Cleanup and Convenience Functions
Fixing problems with the way xml_get_subelement_textsingleton was returning data in cases where there's a mixed model underneath.  Need to be careful whether to call and ask for tail depending on model.  Should still watch this more to see if it's doing the right thing in each case we use it.