Update solrXMLPEPWebLoad.py
Changed the way it checks for modified files...was in MySQL records but set it to use Solr file date instead because it makes less sense to have one mySQL database per solr database.  This way, we can actually check if that version of the file is in Solr.  (You don't know by checking mysql, especially if the database/solr relation isn't one-to-one.)

Also: Added new options:
--before : looks at file date, was it created before this date (use YYYY-MM-DD format)
--after : looks at file date, was it created after this date (use YYYY-MM-DD format)
--reloadbefore : looks at updated (load) date IN SOLR, was it loaded into solr before this date (use YYYY-MM-DD format)
--reloadafter : looks at updated (load) date IN SOLR, was it loaded into solr after this date (use YYYY-MM-DD format)