Update for AWS load
Final changes to 5/31 install to AWS.  Had to add back in the sys.path appends mainly.  Removed the import for solrpy from main.py, since it's not actually used in that module.  Now loaded on AWS as 0311 application (an unfortunate naming there, since I can't update it and change the name, without starting a new instance.)