Updates to schema, the loader program, and the server
Fixed bib fields (in this core) for search or faceting only -- they don't show up in return results, but you can see values in schema analysis or facets.  Fixed loader to normalize these for faceting: removed punctuation and set to lower case.

Abstract return only when requested in search, and no longer fills in the document field with the abstract, that was a convenience so apps could ignore abstract, but adds lots of data. (There was an error, so if the document was accesslimited, abstract always was returned,  even if not requested in the search.)

XML return of the abstract is now supported in search, and a field is added to allow that. The XML is a fragment...you can add your own front matter with the documentInfoXML field  which is currently always returned. Very easy to turn that and the abstract into a  pepkbd3 compliant document with the returned data, and I think it's better to give the  client the choice here.)