Create requirements.txt
Set up environment and minimum requirements to run independently.  However, note that it will still look for localsecrets.py.  If it's not in ../config, then it a copy of that needs to be in the local folder.

Many of the libs are only required if a config requiring tunneling is selected in localsecrets.