First version of the new (derived from solrXMLPEPWebLoad) load program
The files updated here all are concerned with the new AWS S3 loader.

This load program currently has been tested with AWS S3 only, but its destined to replace solrXMLPEPWebLoad, and thus will run on whichever file system is configured in localsecrets.

It does have fewer options than solrXMLPEPWebLoad, so it remains to be determined if it will be "sufficient" to replace solrXMLPEPWebLoad.  However, if both continue to exist, the common routines will all be in the new opasSolrLoadSupport library, so there won't be redundant code in any case.