2020.0929 alpha release changes
2020.0929

Added highlightlimit to /v2/Database/SearchParagraphs/, @GET /v2/Database/Search/, and @POST /v2/Database/Search/. Set it to 0 to turn off highlight returns, and to a positive number to set the number of returned items. Otherwise, you get the default max (5)
Fixed bug where an error in the downloads endpoint returned documentListItem records to the filename instead of throwing an error (and returning no filename).
Changed client-session and client-id per request to check query parameters and then cookies for client-session and client-id until found (or not). Doc still shows header for this, but it will check the others if not found.
Fixed some new errors in constants (and tests) regarding sort field naming
Fixed issue where latest change in the fs library for opasLoad broke image and PDF downloads!