Moved Code to use PaDS only for docs/login
Lots of changes to tests as a result; took PaDS server results and turned them into Pydantic models to make working with them cleaner.  Tricky part was dealing with logging and session ids from PaDS.  This version passes local tests, but I can't seem to test with the remote server...maybe I need whitelisting again?  So I'm going to try to load this on Stage.