2020.1112.1.Alpha
Change /v2/Documents/Glossary/{termIdentifier}/ to use fields group_name_terms for Group and term_terms for Name type loads respectively.  These are copies of the fields group_name and term, but are of type text_simple rather than sring so that subwords are searched correctly (word search instead of string matching).

This may not have been necessary for Name searches, because terms are not paired/trippled up as they are in the group name.  But can always revert that part.