2020.1124.2.Alpha.PySolr
Version based entirely on PySolr middleware.  New opasPySolrLib module to try and separate out pySolr dependent routines, but because of import issues with modules importing each other, had to move more than I would have liked from OpasQueryHandler, so I may try and adjust that some, as well as try to generalize a bit more where routines don't share the calls to PySolr (mainly because the processing of the return data is different, and that's the majority of the return.

Looks to be working pretty well; passed unit test suite and initial usage with local Gavant client.