2021.01.21.2 - Error recovery
Turned on resolve entities, but there shouldn't be entities to resolve.  The problem was in the DTD where I had turned off charents so datamsng was not supplied.  When there's an entity error, lxml apparently doesn't translate ANY entities, so any file that had this missing entity wasn't having entities converted to UTF8.  recovery in the parser is not set to true, so it doesn't exit, and will try to recover.  And while I have removed/fixed instances with datamsng, I will turn the entity back on in the DTD just in case it ever appears again...that was a hard one to figure out!