2021.0412.1
Changed parsing of source_code field for search--previously, it tried to accept a source_name as well, if the entry didn't match a known source code.  But that was causing problems if the input included something unexpected, like parentheses around a single code (CPS).

So now it only accepts source codes, doesn't check them against known codes and try to "help".  Also, now correctly accepts the list as documented so you don't have to use booleans.