2021.0424.2.Beta.UCWorkaround
This build has a simple workaround for a bug I've found that only happens on AWS.  If you submit a fulltext or smarttext query in all uppercase, the server takes a long time and most of the time times out.  Change one letter to lowercase, and it works fine.  This workaround looks for all uppercase in the fulltext1 and smarttext parameters, and if it's all uppercase, converts it to lowercase, since the search result is exactly the same.

If we can figure this out later we can remove this workaround, but it's functionally the same.  The workaround itself is in check_search_args and marked as such.