Update Schema to Fields in models.py per Pydantic change
Breaking change in Pydantic with 1.8... changed Schema to Field, and no longer offering the equivalency.  Interesting though--seems to be breaking loading AWS but not my local copy, and both appear to be using 1.6.1 as per requirements.txt.  Trying the upgrade from Schema to Fields to see if that works on AWS now (while still using 1.6.1)
#  requirements.txt