opasDataPurger 2022.1115/v1.0.001
New module, currently within the opasDataLoader folder.  This does a cleanup of an existing Solr/SQL database structure for loaded files that no longer exist.  This is necessary, for example, when just doing updates, if there were article replacements where the numbering (articleID) changed, as recently happened in work by Aptara and FA.  Otherwise, the previous versions of articles remain in the database until a full rebuild is done.  Currently, this has to be manually started.