2022.1211/v2.1.199
- Adjust new documents/archival endpoint to use output build rather than input build
- Move default build types to opasConfig.
- Regarding this:
TBD: need to remove the option to specify builds in the loader, since this could make it out of sync with the server, which has no such startup option.
Could also move the rest of the XML options from loaderConfig to opasConfg but not "necessary" as they're only used in opasDataLoader.