2023.0303/v2.2.020 / opasUpdateStat 2023.0304/v1.1.12
An art id with a probability attached must have been put in api_biblioxml
during a manual edit, and this was causing a solr query error when put to solr.  This didn't affect processing overal, but to get rid of this error, adapted ArticleID class to recognize probabilities suffixed via : to a art_id, and make them part of the object info, but not the standardized ID (thus removing the probability from that when found).