OpasDataLoader GC Test Version 2023.0607/v2.1.045
This will print out object and memory every commit period (lowered to 250).  It also manually free's (sets vars to None) some memory after each file (even though they should be freed when set during the next loop.

The list memory is still increasing...not obvious where yet.  But I think GC helps a bit, so this might make it through a large list of files.

We may need to look at increasing the amount of memory available during a build.  I think this explains why the errors are only seen online.

This version will be SLOW with the memory checks.